requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
customtkinter>=5.2.0
darkdetect==0.8.0
idna==3.10
packaging==24.2
pillow>=10.0.0
pycryptodome>=3.17.0
requests>=2.28.0
urllib3==2.3.0
CTkMenuBar>=0.8
CTkToolTip>=0.8
kivy>=2.2.1
PyQt6>=6.5.0
qt-material>=2.14
PyInstaller>=5.10.0